+++ .env.example
@@ -93,8 +93,7 @@
 # EVENT_DATABASE_NAME="madsci_events"
 # EVENT_COLLECTION_NAME="events"
 # EVENT_ALERT_LEVEL=40
-# EVENT_EMAIL_ALERTS=null
-# EVENT_EVENT_HANDLERS=["madsci.event_manager.event_handlers.default_error_handler","madsci.event_manager.event_handlers.default_notification_handler"]
+# EVENT_EVENT_HANDLERS=["madsci.event_manager.event_handlers.default_error_handler.EventHandler","madsci.event_manager.event_handlers.default_notification_handler.EventHandler"]
 # EVENT_RETENTION_ENABLED=false
 # EVENT_SOFT_DELETE_AFTER_DAYS=90
 # EVENT_HARD_DELETE_AFTER_DAYS=365
